Refactor CheckAll to separate component (#2853)
* Refactor Check All to component

* Unused import

* cleanup
