+++ components/form-builder/app/responses/DownloadTable.tsx
@@ -23,11 +23,7 @@ import {
 } from "./DownloadTableReducer";
 import { getDaysPassed } from "@lib/clientHelpers";
 import { Alert } from "@components/globals";
-import {
-  CheckAllIcon,
-  CheckBoxEmptyIcon,
-  CheckIndeterminateIcon,
-} from "@components/form-builder/icons";
+import { CheckAll } from "./CheckAll";
 import { DownloadButton } from "./DownloadButton";
 import { toast } from "../shared";
 import { MoreMenu } from "./MoreMenu";
@@ -68,47 +64,6 @@ export const DownloadTable = ({
   });
 
   const MAX_FILE_DOWNLOADS = responseDownloadLimit;
-
-  enum allCheckedState {
-    NONE = "NONE",
-    SOME = "SOME",
-    ALL = "ALL",
-  }
-
-  /**
-   * Are all items checked, some items checked, or no items checked?
-   * @returns {allCheckedState} The current state of the checked items.
-   */
-  const checkAllStatus = () => {
-    if (tableItems.checkedItems.size === 0) {
-      return allCheckedState.NONE;
-    }
-    if (tableItems.checkedItems.size === tableItems.sortedItems.length) {
-      return allCheckedState.ALL;
-    }
-    return allCheckedState.SOME;
-  };
-
-  /**
-   * Check all items if none are checked, otherwise uncheck all items.
-   */
-  const handleCheckAll = () => {
-    if (checkAllStatus() === allCheckedState.NONE) {
-      vaultSubmissions.forEach((submission) => {
-        tableItemsDispatch({
-          type: TableActions.UPDATE,
-          payload: { item: { name: submission.name, checked: true } },
-        });
-      });
-    } else {
-      tableItems.checkedItems.forEach((_, key) => {
-        tableItemsDispatch({
-          type: TableActions.UPDATE,
-          payload: { item: { name: key, checked: false } },
-        });
-      });
-    }
-  };
 
   const handleChecked = (e: React.ChangeEvent<HTMLInputElement>) => {
     const name = e.target.id;
@@ -196,43 +151,12 @@ export const DownloadTable = ({
         <thead className="border-b-2 border-[#6a6d7b]">
           <tr>
             <th className="p-4 text-center">
-              <span
-                className="cursor-pointer"
-                role="checkbox"
-                aria-checked={
-                  checkAllStatus() === allCheckedState.ALL
-                    ? "true"
-                    : checkAllStatus() === allCheckedState.SOME
-                    ? "mixed"
-                    : "false"
-                }
-                tabIndex={0}
-                onClick={handleCheckAll}
-                onKeyDown={(e) => {
-                  if (e.key === "Enter" || e.key === " ") {
-                    handleCheckAll();
-                  }
-                }}
-              >
-                {checkAllStatus() === allCheckedState.ALL && (
-                  <CheckAllIcon
-                    title={t("downloadResponsesTable.header.deselectAll")}
-                    className="m-auto h-6 w-6"
-                  />
-                )}
-                {checkAllStatus() === allCheckedState.SOME && (
-                  <CheckIndeterminateIcon
-                    title={t("downloadResponsesTable.header.deselectAll")}
-                    className="m-auto h-6 w-6"
-                  />
-                )}
-                {checkAllStatus() === allCheckedState.NONE && (
-                  <CheckBoxEmptyIcon
-                    title={t("downloadResponsesTable.header.selectAll")}
-                    className="m-auto h-6 w-6"
-                  />
-                )}
-              </span>
+              <CheckAll
+                tableItems={tableItems}
+                tableItemsDispatch={tableItemsDispatch}
+                noSelectedItemsError={noSelectedItemsError}
+                setNoSelectedItemsError={setNoSelectedItemsError}
+              />
             </th>
             <th className="p-4 text-left">{t("downloadResponsesTable.header.number")}</th>
             <th className="p-4 text-left">{t("downloadResponsesTable.header.status")}</th>
